+++ cabin-sorting/api/mongodb.js
@@ -2,7 +2,7 @@ const { MongoClient } = require('mongodb')
 
 const uri = process.env.MONGO_PROD_CONNECTION_STRING
 // DONT PUSH THIS AT ALL EVER (password and username must be private)
-const client = new MongoClient('REDACTED', {
+const client = new MongoClient(uri, {
   useNewUrlParser: true,
   useUnifiedTopology: true
 })
