+++ cabin-sorting/package.json
@@ -1,23 +1,24 @@
 {
-  "name": "cabin-sorting",
-  "version": "1.0.0",
-  "main": "hackerSortingAlgo.js",
-  "license": "MIT",
-  "dependencies": {
-    "csv-parse": "^5.3.10",
-    "node": "^20.2.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.2.3",
-    "@typescript-eslint/eslint-plugin": "^5.28.0",
-    "@typescript-eslint/parser": "^5.28.0",
-    "eslint": "^8.0.1",
-    "eslint-config-next": "12.1.6",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-n": "^15.0.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.30.0",
-    "eslint-pre-commit-hook": "^2.0.7"
-  }
+    "name": "cabin-sorting",
+    "version": "1.0.0",
+    "main": "hackerSortingAlgo.js",
+    "license": "MIT",
+    "dependencies": {
+        "csv-parse": "^5.3.10",
+        "mongodb": "^5.6.0",
+        "node": "^20.2.0"
+    },
+    "devDependencies": {
+        "@types/node": "^20.2.3",
+        "@typescript-eslint/eslint-plugin": "^5.28.0",
+        "@typescript-eslint/parser": "^5.28.0",
+        "eslint": "^8.0.1",
+        "eslint-config-next": "12.1.6",
+        "eslint-config-standard": "^17.0.0",
+        "eslint-plugin-import": "^2.25.2",
+        "eslint-plugin-n": "^15.0.0",
+        "eslint-plugin-promise": "^6.0.0",
+        "eslint-plugin-react": "^7.30.0",
+        "eslint-pre-commit-hook": "^2.0.7"
+    }
 }
